pyproject.toml
[tool.poetry]
name = "Liz2Bird"
version = "0.1.0"
description = ""
authors = ["EmptyChan <chen19941018@live.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
grequests = "*"
chardet = "*"
logme = "*"
pymongo = "*"
redis = "*"
mmh3 = "*"
colorama = "*"
wrapt = "1.11.1"
aiohttp = "*"
apscheduler = "*"
pymysql = "*"
aiomysql = "*"
retrying = "*"
elasticsearch = "*"
lxml = "*"
python-dateutil = "*"
envparse = "*"
html5lib = "*"
bs4 = "*"
cchardet = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
